Enum eventstore_tcp::adapted::AdaptedMessage

pub enum AdaptedMessage<'a> {
    HeartbeatRequest,
    HeartbeatResponse,
    Ping,
    Pong,
    WriteEvents(WriteEvents<'a>),
    WriteEventsCompleted(Result<WriteEventsCompleted, WriteEventsFailure>),
    ReadEvent(ReadEvent<'a>),
    ReadEventCompleted(Result<ResolvedIndexedEvent<'a>, ReadEventError<'a>>),
    ReadStreamEvents(ReadDirection, ReadStreamEvents<'a>),
    ReadStreamEventsCompleted(ReadDirection, Result<ReadStreamCompleted<'a>, ReadStreamError<'a>>),
    ReadAllEvents(ReadDirection, ReadAllEvents),
    ReadAllEventsCompleted(ReadDirection, Result<ReadAllCompleted<'a>, ReadAllError<'a>>),
    BadRequest(BadRequestMessage<'a>),
    NotHandled(NotHandledInfo<'a>),
    Authenticate,
    Authenticated,
    NotAuthenticated(NotAuthenticatedMessage<'a>),
}

Enumeration of converted messages for more oxidized API.

Variants
HeartbeatRequest

Requests heartbeat from the other side. Unsure if clients or server sends these.

HeartbeatResponse

Response to a heartbeat request.

Ping

Ping request, similar to heartbeat.

Pong

Ping response.

WriteEvents(WriteEvents<'a>)

Append to stream request

WriteEventsCompleted(Result<WriteEventsCompleted, WriteEventsFailure>)

Append to stream response, which can fail for a number of reasons

ReadEvent(ReadEvent<'a>)

Request to read a single event from a stream

ReadEventCompleted(Result<ResolvedIndexedEvent<'a>, ReadEventError<'a>>)

Response to a single event read

ReadStreamEvents(ReadDirection, ReadStreamEvents<'a>)

Request to read a stream from a point forward or backward

ReadStreamEventsCompleted(ReadDirection, Result<ReadStreamCompleted<'a>, ReadStreamError<'a>>)

Response to a stream read in given direction

ReadAllEvents(ReadDirection, ReadAllEvents)

Request to read a stream of all events from a position forward or backward

ReadAllEventsCompleted(ReadDirection, Result<ReadAllCompleted<'a>, ReadAllError<'a>>)

Response to a read all in given direction

BadRequest(BadRequestMessage<'a>)

Request was not understood. Please open an issue!

NotHandled(NotHandledInfo<'a>)

Correlated request was not handled. This is the likely response to requests where require_master is true, but the connected endpoint is not master and cannot reach it.

Authenticate

Request to authenticate attached credentials.

Authenticated

Positive authentication response. The credentials used to Authenticate previously can be used in successive requests.

NotAuthenticated(NotAuthenticatedMessage<'a>)

Negative authentication response, or response to any sent request for which used authentication was not accepted.

Methods

impl<'a> AdaptedMessage<'a>

fn as_raw<'b>(&'b self) -> RawMessage<'b>

Converts the adapted back to raw for encoding.