Module eventstore_tcp::builder

Builders to help building requests.

Every top-level builder exposes at least method build_package which will return a ready-to-send Package but will panic if one or more of the required properties is not set. Currently the builders are implemented so that it should support chaining.

On build_package the contents are not cloned but are moved into the returned Package which is not apparent from the use of &mut self.

UsernamePassword can be used to authenticate specific requests. Specifying credentials for every request is currently the only supported way to authenticate.

Any Option<Uuid> can always be left unset, in which case Uuid::new_v4() (random Uuid) is generated.

Structs

Builder
Factory factory for creating builders.

DeleteStreamBuilder
Builder for DeleteStream.

NewEventBuilder
Builder for specifying an event when using WriteEventsBuilder.

ReadAllEventsBuilder
Builder for ReadAllEvents.

ReadEventBuilder
Builder for a single event read request ReadEvent.

ReadStreamEventsBuilder
Builds a package for reading multiple events from a stream in either ReadDirection.

SimpleBuilder
Builder for messages without any additional contents.

WriteEventsBuilder
Builder for WriteEvents which allows writing multiple events to a stream.