Struct eventstore_tcp::builder::Builder

pub struct Builder;

Factory factory for creating builders.

Methods

impl Builder

fn ping() -> SimpleBuilder

Simple builder for a Ping message, that has no other data.

fn authenticate() -> SimpleBuilder

Simple builder for an Authenticate message, that has no other data (credentials are passed to the build_package method).

fn write_events() -> WriteEventsBuilder

Builder for WriteEvents, which allows writing multiple events to a stream, with expected current version of the stream.

fn read_event() -> ReadEventBuilder

Builder for ReadEvent which allows reading a single event off a stream.

fn read_stream_events() -> ReadStreamEventsBuilder

Builder for ReadStreamEvents which allows reading multiple events off a stream either forwards or backwards.

fn read_all_events() -> ReadAllEventsBuilder

Builder for ReadAllEevents which allows reading off a stream of all events in the database.