Struct eventstore_tcp::builder::ReadAllEventsBuilder

pub struct ReadAllEventsBuilder { /* fields omitted */ }

Builder for ReadAllEvents.

Methods

impl ReadAllEventsBuilder

fn direction<D: Into<ReadDirection>>(&mut self, dir: D) -> &mut Self

Set the read direction (required).

fn positions<N: Into<LogPosition>, M: Into<LogPosition>>(&mut self,
commit: N,
prepare: M)
-> &mut Self

Sets the positions to read from. These are easiest to acquire from previous ReadAllSuccess responses, likely persisted somewhere between reads.

fn max_count(&mut self, count: u8) -> &mut Self

Sets the maximum number of events to read (required). Panics if argument is zero. u8 is used as larger batches should be paged. At the moment maximum buffer requirement even for 255 events is 255*16MiB > 4000MB.

fn resolve_link_tos(&mut self, resolve: bool) -> &mut Self

Whether or not the server should resolve links found in the stream to events of other streams. Defaults to true.

fn require_master(&mut self, require: bool) -> &mut Self

Should the server only handle the request if it is the cluster master. Note that while only the master server can accept writes, other cluster members can forward requests to the master.

Defaults to false.

fn build_package(&mut self,
authentication: Option<UsernamePassword>,
correlation_id: Option<Uuid>)
-> Package

Build a package. Will panic if required values are not set. Values of this builder will be moved into the package.