Enum eventstore_tcp::Message

pub enum Message {
    HeartbeatRequest,
    HeartbeatResponse,
    Ping,
    Pong,
    WriteEvents(WriteEvents<'static>),
    WriteEventsCompleted(Result<WriteEventsCompleted, WriteEventsFailure>),
    ReadEvent(ReadEvent<'static>),
    ReadEventCompleted(Result<ResolvedIndexedEvent<'static>, ReadEventFailure>),
    ReadStreamEvents(ReadDirection, ReadStreamEvents<'static>),
    ReadStreamEventsCompleted(ReadDirection, Result<ReadStreamSuccess, ReadStreamFailure>),
    BadRequest(Option<String>),
    NotHandled(NotHandledReason, Option<MasterInfo<'static>>),
    Authenticate,
    Authenticated,
    NotAuthenticated,
}

Enumeration of currently supported messages. The plan is to include every defined message while trying to decode responses into Result alike messages, such as ReadEventCompleted.

Variants
HeartbeatRequest

Requests heartbeat from the other side. Unsure if clients or server sends these.

HeartbeatResponse

Response to a heartbeat request.

Ping

Ping request, similar to heartbeat.

Pong

Ping response.

WriteEvents(WriteEvents<'static>)

Append to stream request

WriteEventsCompleted(Result<WriteEventsCompleted, WriteEventsFailure>)

Append to stream response, which can fail for a number of reasons

ReadEvent(ReadEvent<'static>)

Request to read a single event from a stream

ReadEventCompleted(Result<ResolvedIndexedEvent<'static>, ReadEventFailure>)

Response to a single event read

ReadStreamEvents(ReadDirection, ReadStreamEvents<'static>)

Request to read a stream from a point forward or backward

ReadStreamEventsCompleted(ReadDirection, Result<ReadStreamSuccess, ReadStreamFailure>)

Response to a stream read in given direction

BadRequest(Option<String>)

Request was not understood. Please open an issue!

NotHandled(NotHandledReason, Option<MasterInfo<'static>>)

Correlated request was not handled. This is the likely response to requests where require_master is true, but the connected endpoint is not master and cannot reach it.

Authenticate

Request to authenticate attached credentials.

Authenticated

Positive authentication response. The credentials used to Authenticate previously can be used in successive requests.

NotAuthenticated

Negative authentication response, or response to any sent request for which used authentication was not accepted.