Struct eventstore_tcp::package::Package

pub struct Package {
    pub authentication: Option<UsernamePassword>,
    pub correlation_id: Uuid,
    pub message: MessageContainer<'static>,
}

Frame in the protocol. On the wire, packets are embedded in frames with length prefix and suffix.

Fields
authentication: Option<UsernamePassword>

Possible authentication data included in the packet. Some and None values of this will be used to generate corresponding TcpFlags first bit.

correlation_id: Uuid

Before sending an request to the server client generates a new random UUID using uuid::Uuid::new_v4() and later server will respond using the same correlation_id.

message: MessageContainer<'static>

Enumeration of possible messages.