Module eventstore_tcp::raw

Raw module contains the enumeration RawMessage and raw decoding and encoding functionality. There should not be need to handle RawMessage values directly but if there is ever a bug, using the raw messages should still work.

Reexports

pub use self::client_messages::EventRecord;

pub use self::client_messages::WriteEvents;

pub use self::client_messages::WriteEventsCompleted;

pub use self::client_messages::ReadEvent;

pub use self::client_messages::ReadEventCompleted;

pub use self::client_messages::ReadStreamEvents;

pub use self::client_messages::ReadStreamEventsCompleted;

pub use self::client_messages::ReadAllEvents;

pub use self::client_messages::ReadAllEventsCompleted;

pub use self::client_messages::NotHandled;

pub use self::client_messages::DeleteStream;

pub use self::client_messages::DeleteStreamCompleted;

pub use self::client_messages::OperationResult;

Modules

client_messages
Automatically generated rust module for 'ClientMessageDtos.proto' file

Structs

BadRequestPayload
Newtype for an arbitary BadRequest "info", which could be UTF8 string.

NotAuthenticatedPayload
Newtype for an arbitary NotAuthenticated "info", which could be UTF8 string.

Enums

RawMessage
Enumeration much like the adapted::AdaptedMessage for all the messages in the protocol.