Enum eventstore_tcp::raw::RawMessage

pub enum RawMessage<'a> {
    HeartbeatRequest,
    HeartbeatResponse,
    Ping,
    Pong,
    WriteEvents(WriteEvents<'a>),
    WriteEventsCompleted(WriteEventsCompleted<'a>),
    DeleteStream(DeleteStream<'a>),
    DeleteStreamCompleted(DeleteStreamCompleted<'a>),
    ReadEvent(ReadEvent<'a>),
    ReadEventCompleted(ReadEventCompleted<'a>),
    ReadStreamEvents(ReadDirection, ReadStreamEvents<'a>),
    ReadStreamEventsCompleted(ReadDirection, ReadStreamEventsCompleted<'a>),
    ReadAllEvents(ReadDirection, ReadAllEvents),
    ReadAllEventsCompleted(ReadDirection, ReadAllEventsCompleted<'a>),
    BadRequest(BadRequestPayload<'a>),
    NotHandled(NotHandled<'a>),
    Authenticate,
    Authenticated,
    NotAuthenticated(NotAuthenticatedPayload<'a>),
    Unsupported(u8, Cow<'a, [u8]>),
}

Enumeration much like the adapted::AdaptedMessage for all the messages in the protocol.

Variants
HeartbeatRequest

Requests heartbeat from the other side. Unsure if clients or server sends these.

HeartbeatResponse

Response to a heartbeat request.

Ping

Ping request, similar to heartbeat.

Pong

Ping response.

WriteEvents(WriteEvents<'a>)

Append to stream request

WriteEventsCompleted(WriteEventsCompleted<'a>)

Append to stream response, which can fail for a number of reasons

DeleteStream(DeleteStream<'a>)

Request to delete a stream

DeleteStreamCompleted(DeleteStreamCompleted<'a>)

Response to previous stream deletion request

ReadEvent(ReadEvent<'a>)

Request to read a single event from a stream

ReadEventCompleted(ReadEventCompleted<'a>)

Response to a single event read

ReadStreamEvents(ReadDirection, ReadStreamEvents<'a>)

Request to read a stream from a point forward or backward

ReadStreamEventsCompleted(ReadDirection, ReadStreamEventsCompleted<'a>)

Response to a stream read in given direction

ReadAllEvents(ReadDirection, ReadAllEvents)

Request to read a stream of all events from a position forward or backward

ReadAllEventsCompleted(ReadDirection, ReadAllEventsCompleted<'a>)

Response to a read all in given direction

BadRequest(BadRequestPayload<'a>)

Request was not understood. Please open an issue!

NotHandled(NotHandled<'a>)

Correlated request was not handled. This is the likely response to requests where require_master is true, but the connected endpoint is not master and cannot reach it.

Authenticate

Request to authenticate attached credentials.

Authenticated

Positive authentication response. The credentials used to Authenticate previously can be used in successive requests.

NotAuthenticated(NotAuthenticatedPayload<'a>)

Negative authentication response, or response to any sent request for which used authentication was not accepted. May contain a reason.

Unsupported(u8, Cow<'a, [u8]>)

Placeholder for a discriminator and the undecoded bytes

Methods

impl<'a> RawMessage<'a>

fn into_owned(self) -> RawMessage<'static>

Turns possibly borrowed value of self into one that owns all of it's data.

fn decode(discriminator: u8, buf: &'a [u8]) -> Result<RawMessage<'a>>

Decodes the message from the buffer without any cloning.

fn encode<W: Write>(&self, w: &mut W) -> Result<()>

Encodes the message into the given writer.

fn discriminator(&self) -> u8

Returns the protocol discriminator value for the variant