Struct eventstore_tcp::EventStoreClient

pub struct EventStoreClient { /* fields omitted */ }

tokio_service::Service implementation of the client.

Methods

impl EventStoreClient

fn connect(addr: &SocketAddr,
handle: &Handle)
-> Box<Future<Item=Self, Error=Error>>

Connect to an EventStore database listening at given addr using the given tokio::reactor::Cores handle. Returns a future representing the client which can be used to send and receive Package values.