Struct eventstore_tcp::ReadStreamSuccess

pub struct ReadStreamSuccess {
    pub events: Vec<ResolvedIndexedEvent<'static>>,
    pub next_page: Option<EventNumber>,
    pub last_event_number: StreamVersion,
    pub end_of_stream: bool,
    // some fields omitted
}

Successful response to a Message::ReadStreamEvents.

Fields
events: Vec<ResolvedIndexedEvent<'static>>

The actual events returned by the server. Subject to resolve_link_tos setting on the read request.

next_page: Option<EventNumber>

EventNumber for a query for the next page in the same direction, None if start has been reached when reading backwards. When reading forwards, this will never be None as new events might have appeared while receiving this response.

last_event_number: StreamVersion

Event number of the last event

end_of_stream: bool

Has the end of the stream been reached (or could more events be read immediatedly)